+++ packages/vue/src/server.ts
@@ -1,0 +1,1 @@
+export { default as default } from '@wreathe-js/core/server'