+++ packages/vue/CHANGELOG.md
@@ -7,4 +7,17 @@ project adheres to [Semantic Versioning](https://semver.org/).
 
 ## Unreleased
 
-- Initial release
+- initial release based on [@inertiajs/vue](https://github.com/inertiajs/inertia) (v1.0.0-beta.2)
+
+### Refactoring
+
+- convert adapter to TypeScript
+- change namespace `Inertia/inertia` to `Wreathe/wreathe`
+- remove `microbundle` dependency
+
+### Other
+
+- add `unbuild` dependency
+- add `@types/lodash.clonedeep` dependency
+- add `@types/lodash.isequal` dependency
+- update `vue` dependency to 3.2.45